Rename directories to be all camelCase (#19)
We'd like to have consistency within the directories.

We had some directories that were PascalCase (like
`src/components/ui/GetInvolved`), and some that were camelCase (like
`src/components/Footer`). We'd to be consistent with the casing of
directories, so we convert the all to camelCase.

For more information, see this PR:
https://github.com/Say-Their-Name/web/pull/16.
