+++ src/components/footer/style.js
