+++ src/components/ui/mediaList/MediaList.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import Media from '../Media/Media';
+import Media from '../media/Media';
 import Container from '../../Container';
 import StyledMediaList from './styles';
